+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [1.60.0] (2026-06-23)
 
 Update for Patch 7.51 HotFix 2.
 
@@ -957,7 +957,8 @@ Updated CS for 7.05hf1
 
 First release! 🥳
 
-[unreleased]: https://github.com/Haselnussbomber/HaselDebug/compare/v1.59.4...main
+[unreleased]: https://github.com/Haselnussbomber/HaselDebug/compare/v1.60.0...main
+[1.60.0]: https://github.com/Haselnussbomber/HaselDebug/compare/v1.59.4...v1.60.0
 [1.59.4]: https://github.com/Haselnussbomber/HaselDebug/compare/v1.59.3...v1.59.4
 [1.59.3]: https://github.com/Haselnussbomber/HaselDebug/compare/v1.59.2...v1.59.3
 [1.59.2]: https://github.com/Haselnussbomber/HaselDebug/compare/v1.59.1...v1.59.2